Errors with apostrophe handling
Versions of scala-xml prior to 1.1.0 have a bug where &apos; and ' do
not generate events when parsing. As a result, Daffodil has problems
with unparsing apostrophes (bug #1420) and with the TDML runner eating
them (bug #1928) which cannot be fixed without updating scala-xml.

This commit updates the scala-xml dependency to 1.1.0, but this DOES
NOT fix the issues, so the tests remain disabled. Also bump SBT to
1.1.2 so that correct scala-xml version is picked up for classpaths.

DAFFODIL-1420, DAFFODIL-1928